FormModal moved to separate file to clean up map.js
