+++ client/src/App.css
@@ -116,17 +116,21 @@ input {
   height: 20vw;
   background-color: #333;
   border-radius: 20px;
+  display: flex;
+  flex-direction: column;
 }
 .modal-card {
   /* width: 15vw; */
   width: 100%;
+  height: 100%;
   margin: 0;
   box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
   border-radius: 5px;
   display: flex;
   flex-direction: column;
-  justify-content: center;
-  align-content: space-between;
+  justify-content: space-between;
+  flex: 1;
+
 }
 
 .modal-card-head {
@@ -148,6 +152,8 @@ input {
 
 .modal-card-body {
   padding: 1rem;
+  flex: 1;
+
 }
 
 .delete {
